luci-base: uci: (APP-2203) stop uci.get modifying internal state
The 'v' object is supposed to reflect the underlying uci state,
but if you get an entire section it applies the current changes
to that section. This confuses subsequent operations, most notably
causing spurious deletes if you go through a set/get/delete
sequence.

Basically the only change here is to take a copy of the section
before applying changes.

Approved-by: Arien Judge
Approved-by: Evan Benn