luci-base: (APP-2561) allow calling Section.formvalue in all situations
If you called Section.formvalue before the map was rendered, it
fell back to cfgvalue (i.e. the uci value), and _after_ it was
rendered it used the formvalue. However, while rendering was happening
(e.g. in Option.load) this.map.root was constructed but not populated.
This catches this situation and uses cfgvalue.

Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn