luci-base: (APP-2761) don't render until luci-loaded is done
Because the setupDOM/initDOM methods do strange things related
to the old server rendered templates
(like remove all the elements currently hidden by dependencies...),
we want to be sure that these have finished before the view
itself renders.

Strictly, it would probably be better to put this delay after
the render and before we insert the DOM, but luci-loaded should
be reasonably quick, and in case anyone is particularly creative
in the render I feel it's safer to do this in the load.


Approved-by: Arien Judge
Approved-by: Evan Benn