luci-mod-status: (APP-3339, PR #113) luci logs page fast scroll on varying theme
Themes which have a side menu bar won't have the entire
document by scrollable, but only the main content div.
A simple option for logs scrolling is therefore
just to scroll to the location of the other button.

This is the same fix that's in upstream (but in a single commit); we can drop it next time we rebase.

Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn