luci-base/luci-mod-network: add Morse HaLow support
This adds basic Morse HaLow frequency support to LuCI.
It's a roll-up of a large number of smaller commits
(i.e. I am not the real author). These are squashed
so it's easier to apply for future rebases.

Currently it relies on having the morse-regdb package
(as not all the information available via here is available
via the driver or the current wireless regdb).