luci-app-statistics: (APP-3253, PR #126) render all graphs on refresh if no summary
Previously, if there would usually be a summary page for the data
but this was not happening because there was only a single instance
(e.g. a single wifi interface), the refresh would mistakenly
request the summary rather than the existing data.

i.e. graphs would mysteriously disappear on refresh.

Raised with upstream: https://github.com/openwrt/luci/pull/7344


Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn