luci-base: (APP-3592, PR #129) fix repeated load of country list
load() in an AbstractValue, which is usually used to load the
underlying value from UCI, is a bad place to load the possible
form values as it may be called more than once. i.e. if you
save the form, it's called again to refresh the contents.

However, it appears there's no _nice_ place to do this,
as the possible values are dependent on uci values for
the section. So we copy the hacky solution from:

https://github.com/openwrt/luci/pull/4286/files

(note that this widget is in the same file as this one;
there are other widgets in the file that are still broken, but
you don't see the issue because they're only use in modals
that disappear on save)

Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn