wireless.js: APP-3673: (PR #133) Remove '1' option from mesh_rssi_threshold
mesh_rssi_threshold was copied from the other wireless config section.
However morse mesh_rssi_threshold is a mesh11sd setting, while the
source is using mesh_rssi_threshold as a hostap setting. hostap has
the logic to use '1' as a 'do not change' setting. mesh11sd does not
have this logic. Therefore remove the option to set the value 1.

The special meaning of '0' comes from the kernel.

Also make it an integer.

Approved-by: James Haggerty
Approved-by: Sophronia Koilpillai