luci-base: (APP-3796, PR #144) provide previousValue in onchange
This allows an option which has an onchange callback to consider
the prior value when deciding what it should do.

When initially rendering the widget previous value wasn't set.
This didn't really matter for upstream, because previousValue was only used
for comparison, but now we're passing previousValue to onchange...

I've broken my rule here about keeping with upstream style, because
it's a bit of promise spaghetti if you want to pass the original cfg_value
through.

At the same time, I've had to modify renderUpdate to also deal with this
new argument to renderFrame apppropriately, and since I was async-ifying
render I thought I may as well do the same to renderUpdate
(since we actually own that completely).

Approved-by: Sophronia Koilpillai
Approved-by: Lyall Beveridge