luci-mod-network (APP-3939, APP-3899, PR #150) Clean up DH Group situation
This goes back on most of the changes in 76e20b82f773e2c85c4f11509ba9dbea99dbc8f1,
and instead restricts groups to (a) those that seem to mostly work (i.e. hostapd supports)
and (b) reinstates the ability to select multiple groups (ListValue -> MultiValue).

We also remove the MODP groups from OWE (since they don't work - some complaint
about being unable to build the IE for the association frame)
and make sure that this is only accessible for the morse chip.
I've also move it to the bottom, since users are very unlikely to want to modify this.

Note that it's not worth investing much time here: normal OpenWrt doesn't support
it since everyone just uses 19, and I believe this was only 'added' for us
as it was a WFA test requirement. I only haven't removed it entirely because
of the sunk cost fallacy and lack of enthusiasm for talking about DH Groups.


Approved-by: Sophronia Koilpillai
Approved-by: Lyall Beveridge