luci-base/luci-mod-network: (APP-1976, PR #153) filter primary chan index based on freqlist
Previously, we either let you select anything OR forbid based on pre-calculation in
morse-regdb. Now we figure out whether the primary channel is allowed before letting
you select an index.

Logic here is inspired by add_primary_channels.py in morse-regdb, which I'll
remove in a separate PR.

Note that to achieve this we move the primary channel logic into the
WifiFrequencyWidget, which IMO makes it easier to see what's going on
for the user and slightly simplifies the twisty maze of code here.

Approved-by: Sophronia Koilpillai
Approved-by: Lyall Beveridge