luci-base: (APP-4133, PR #158) correctly select channels on country change
Previously this was relying on a dispatched event, and when I added the
primary channel support I 'copied' the approach of the below
functions (which worked most of the time...).

To avoid confusion, I've removed the dispatch and instead
manually called the appropriate toggle functions. This is
somewhat inconsistent with upstream, but because we have
so _many_ chained updates I think it's worth it.


Approved-by: Sophronia Koilpillai
Approved-by: Andrew Dodd