luci-base: (APP-4138) align formvalue/cfgvalue in wifi widget
formvalue and cfgvalue should generate equivalent data, as this
is what determines whether the value is written.

Previously, the value was _always_ being written even on no change
due to issues here. This wasn't very important, but in order
to support removing prim_channel only on change I needed to fix
this properly.