luci-base: (APP-4138) remove primary channel + width on change
Previously, we only touched the s1g primary channel if we had
the option enabled to show the channel/width (primChanSelect = true).
However, this meant that if you changed the channel when
primChanSelect = false, the old channel/width remained.

Now we write in _all_ situations, relying on the code in form.js
not to perform a write if the formvalue/cfgvalue are identical.
i.e. only write if the channel (or some other param) changes.