luci-base/luci-mod-network: (APP-4170, PR #160) show all supported encryptions in adv wireless config
This normalises the morse 'stuff' so it works the same way as the OpenWrt
hostap/wpa_supplicant feature detection.

Why do we want this? We were 'removing' modes that worked (like psk2/sae-mixed)
that were used in prplmesh, which meant that if you went to this page
and saved it would be destructive of your config.

The downside to doing this is that:
- we no longer have the 'only show the valid options' in mesh mode
  (but use standard OpenWrt's mechanism of marking them invalid)
- we have to filter out a few options that we really don't want
  manually (e.g. wpa-eap/wpa2-eap/wpa-psk)
- this requires a patch to hostapd/wpa_supplicant to do feature
  detection


Approved-by: Sophronia Koilpillai
Approved-by: Prabha Murugesan