luci-base: (APP-4330, PR #165) improve modal escape handling
Confusingly, each modal gets to choose how to build itself, but
there's generic 'Escape' handling which tries to click a button
(based on CSS classes intended for styling...).

What this means is that if, inside the modal, you have another
right > btn thing (e.g. inside the cbi-filebrowser), the Escape
handling will find it before it finds the Dismiss button.

This seems fundamentally broken to me, because each modal
gets to choose everything about itself BUT the escape handling
assumes that the modal is constructed in a certain way. For now,
we workaround the issue here by only clicking a button
in the .right immediately under the modal div (rather
than one anywhere in the DOM).


Approved-by: Prabha Murugesan
Approved-by: Lyall Beveridge