luci-mod-network: (APP-4943, PR #186) use placeholder instead of this.defaults for beacon_int
Using this.defaults meant that the beacon_int was persisted all the time,
causing unnecessary changes. Instead, we should use the placeholder
to accurately reflect the state of the empty configuration.

This is apparently the way other dynamic placeholders are set in luci,
even though it's very ugly (hooking into the fact that the validate
method is called on any change for all active fields).


Approved-by: Prabha Murugesan
Approved-by: Evan Benn