luci-mod-base: (APP-4880, PR #182) support auto chans in halow chan selector
This allows you to select auto (i.e. ACS) as the channel,
and also forces EU/GB to only show auto.

If auto is set, it also writes the bandwidth and makes
sure you can't set the primary channel index (as
it's impossible to know which primary channel
indexes are valid, we pass this problem down
and let it fail there...).

* luci-base: (APP-4880) retain s1g-primary-channel if no change

This requires us to predict the bandwidth of the selected
channel in cfgvalue() so we make sure it's the same as
formvalue().

* luci-base: (APP-4880) rename bandwidth to s1g_chanbw

This makes us closer to the existing chanbw naming.

* luci-base: (APP-4880) improve unknown channel selection logic

Prefer middle channel over auto when changing bw/country
and presenting a new option.

* luci-base: (APP-4880) make disableACS use clearer

Approved-by: Prabha Murugesan
Approved-by: Andrew Dodd